refactor: move use solana into lib directory
